Fix some codes to fit travis CI & sonarcloud
