+++ sonar-project.properties
@@ -3,7 +3,7 @@ sonar.organization=swsnu
 sonar.projectName=swpp2020-team15
 sonar.sources=stockin/frontend/src, stockin/backend
 sonar.tests=stockin/frontend/src, stockin/backend
-sonar.exclusions=stockin/frontend/src/index.js, stockin/frontend/src/api/index.js,stockin/frontend/src/test-utils/mocks.js, stockin/frontend/src/serviceWorker.js, stockin/frontend/src/setupTests.js, stockin/backend/**/migrations/*, stockin/backend/manage.py, stockin/backend/stockin/asgi.py, stockin/backend/stockin/settings.py, stockin/backend/stockin/urls.py, stockin/backend/stockin/wsgi.py, stockin/backend/core/crawlers/*, stockin/backend/core/ML/*
+sonar.exclusions=stockin/frontend/src/index.js, stockin/frontend/src/api/index.js,stockin/frontend/src/test-utils/mocks.js, stockin/frontend/src/serviceWorker.js, stockin/frontend/src/setupTests.js, stockin/backend/**/migrations/*, stockin/backend/manage.py, stockin/backend/stockin/static/**/*, stockin/backend/stockin/asgi.py, stockin/backend/stockin/settings.py, stockin/backend/stockin/urls.py, stockin/backend/stockin/wsgi.py, stockin/backend/core/crawlers/*, stockin/backend/core/ML/*
 sonar.test.inclusions=stockin/frontend/src/**/*.test.js, stockin/backend/**/tests_*.py
 sonar.javascript.lcov.reportPaths=stockin/frontend/coverage/lcov.info
 sonar.python.coverage.reportPaths=stockin/backend/coverage.xml
