+++ stockin/backend/core/views/groups.py
@@ -11,6 +11,7 @@ import json, os, csv
 
 from core.models import Group, Stock, FinancialStat
 from core.crawlers.preprocessors.score import base_score
+from core.views.index import getFSInfo
  
 def group_list_and_create(request):
     if request.method == 'GET':
@@ -29,55 +30,7 @@ def group_list_and_create(request):
             
             for stock in stock_list :
                 fs_stock = FinancialStat.objects.filter(stock_id=stock.id)
-                op = ['','','','','']
-                try:
-                    op[4] = fs_stock.get(quarter='20년 6월').operatingProfit
-                except:
-                    pass
-                try:
-                    op[3] = fs_stock.get(quarter='20년 3월').operatingProfit
-                except:
-                    pass
-                try:
-                    op[2] = fs_stock.get(quarter='19년 12월').operatingProfit
-                except:
-                    pass
-                try:
-                    op[1] = fs_stock.get(quarter='19년 6월').operatingProfit
-                except:
-                    pass
-                try:
-                    op[0] = fs_stock.get(quarter='19년 3월').operatingProfit
-                except:
-                    pass
-                for i in range(5):
-                    if op[i] == '-':
-                        op[i] = ''
-                # liability rate
-
-                # operatingCashflow
-                script_dir = os.path.dirname(__file__) #<-- absolute dir the script is in
-                rel_path = "../crawlers/data/Cash_Flow.csv"
-                abs_file_path = os.path.join(script_dir, rel_path)
-                f = open(abs_file_path, 'r', encoding='utf-8')
-                rdr = csv.reader(f)
-                operatingCashflow = []
-                for line in rdr:
-                    if line[0] == stock.title and line[1] == '18년 12월':
-                        if line[2] == ' ':
-                            operatingCashflow.append('')
-                        else:
-                            operatingCashflow.append(float(str(line[2]).replace(',','')))
-                    elif line[0] == stock.title and line[1] == '19년 12월':
-                        if line[2] == ' ':
-                            operatingCashflow.append('')
-                        else:
-                            operatingCashflow.append(float(str(line[2]).replace(',','')))
-                        break
-                f.close()  
-
-                # response = {'score' : score, 'status': if score is None, operatingProfitNotEnough 1, operatingCashflowNotEnough 2, Both 3}
-                fs_score = base_score(op, operatingCashflow, stock.debtRatio, stock.crawledPER, stock.crawledPERAvg, stock.operatingMarginRate, stock.operatingMarginRateAvg)
+                fs_score = getFSInfo(stock, fs_stock)
                 stock_info_dict = {
                     'id' : stock.id,
                     'title' : stock.title,
@@ -252,8 +205,3 @@ def group_stock_delete(request, group_id='', stock_id=''):
                 
     else :
         return HttpResponseNotAllowed(['DELETE'])
-
-
-
-
-
